XmlDoc field added to class
Interesting class: In the process of creating simple classes with clear
responsibilities, I was relying on the XmlDoc required to make the
XPathQuery being in the scope of the Query.

This will not necessarily be the case, because I want to create a
ResultSet form an XPathQuery factory method, which means the class can
be "exported" out of the immediate scope.