Interface requires redesign!
Using std::unique_ptr and wrapping up the xpath classes makes the
interface difficult to use as well as being inefficient.

XPathQuery objects should be copyable and assignable, which means they
need a reference to the XmlDoc, for which they have the xmlPathObject,
which is currently held as a unique_ptr in XPathCtxt, an unwieldly
design.

As an XPathQuery etc. is going to be the standard use it is better to
redesign XmlDoc to hold a shared_ptr and propogate this change
throughout the classes that need access to an xmlDoc.